+++ src/mlpack/bindings/R/mlpack/src/r_util.cpp
@@ -153,10 +153,34 @@ void IO_SetParamMatWithInfo(const std::string& paramName,
     d.Type(i) = (dimensions[i]) ? data::Datatype::categorical :
         data::Datatype::numeric;
   }
+
+  arma::mat m = paramValue.t();
+
+  // Do we need to find how many categories we have?
+  if (hasCategoricals)
+  {
+    arma::vec maxs = arma::max(paramValue, 1) + 1;
+
+    for (size_t i = 0; i < d.Dimensionality(); ++i)
+    {
+      if (dimensions[i])
+      {
+        // Map the right number of objects.
+        for (size_t j = 0; j < (size_t) maxs[i]; ++j)
+        {
+          std::ostringstream oss;
+          oss << j;
+          d.MapString<double>(oss.str(), i);
+        }
+      }
+    }
+  }
+
   std::get<0>(IO::GetParam<std::tuple<data::DatasetInfo, arma::mat>>(
       paramName)) = std::move(d);
   std::get<1>(IO::GetParam<std::tuple<data::DatasetInfo, arma::mat>>(
-      paramName)) = paramValue.t();
+      paramName)) = std::move(m);
+
   IO::SetPassed(paramName);
 }